xitongchuli

+++ database/migrations/2022_01_21_074957_create_writer_order_table.php
@@ -17,7 +17,7 @@ class CreateWriterOrderTable extends Migration
             $table->bigInteger('writerId')->default(0)->comment('写手唯一ID');
             $table->bigInteger('orderId')->default(0)->comment('订单ID');
             $table->bigInteger('shop_id')->default(0)->comment('商店ID');
-            $table->integer('writerPrice')->default(0)->comment('写手派单价');
+            $table->float('writerPrice')->default(0)->comment('写手派单价');
             $table->integer('wSettleState')->default(2)->comment('写手结算状态 1:已结算,2:未结算,3:暂缓结算');
             $table->integer('compensateState')->default(0)->comment('补偿状态 1:稿费补偿');
             $table->primary(['writerId', 'orderId']);
